Rename: 태그, 컴포넌트명, 폴더명 Dialog로 변경
